pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.1"
description = "Сommand-line interface for comparing two configuration files and displaying the difference"
authors = ["ilitili <momcoveylebozo@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff="gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pydash = "^5.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
